+++ src/VaultSharp/V1/SecretsEngines/TOTP/TOTPCode.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace VaultSharp.V1.SecretsEngines.TOTP
+{
+    /// <summary>
+    /// Represents the TOTP Code.
+    /// </summary>
+    public class TOTPCode
+    {
+        /// <summary>
+        /// Gets or sets the TOTO code.
+        /// </summary>
+        /// <value>
+        /// The TOTP code.
+        /// </value>
+        [JsonProperty("code")]
+        public string Code { get; set; }
+    }
+}